+++ implementations/micrometer-registry-ganglia/src/main/java/io/micrometer/ganglia/GangliaMeterRegistry.java
@@ -160,7 +160,7 @@ public class GangliaMeterRegistry extends StepMeterRegistry {
 
     private void announceSummary(DistributionSummary summary) {
         HistogramSnapshot snapshot = summary.takeSnapshot();
-        announce(summary, snapshot.count(), "count");
+        announce(summary, (double) snapshot.count(), "count");
         announce(summary, snapshot.total(), "sum");
         announce(summary, snapshot.mean(), "avg");
         announce(summary, snapshot.max(), "max");
@@ -168,7 +168,7 @@ public class GangliaMeterRegistry extends StepMeterRegistry {
 
     private void announceTimer(Timer timer) {
         HistogramSnapshot snapshot = timer.takeSnapshot();
-        announce(timer, snapshot.count(), "count");
+        announce(timer, (double) snapshot.count(), "count");
         announce(timer, snapshot.total(getBaseTimeUnit()), "sum");
         announce(timer, snapshot.mean(getBaseTimeUnit()), "avg");
         announce(timer, snapshot.max(getBaseTimeUnit()), "max");
